% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hd_panel_estimator.r
\name{hd_panel_estimator}
\alias{hd_panel_estimator}
\title{Estimate HD panels with IFE}
\usage{
hd_panel_estimator(
  data,
  obs_N,
  obs_T,
  TRUNC = 0.05,
  NFACTORS = NULL,
  variant = "LS",
  penalty_choice = NULL,
  NFOLDS = NULL,
  NLAMBDA = NULL,
  scree_plot = TRUE
)
}
\arguments{
\item{data}{List containing the balanced panel data with data$y containing
the dependent variables and data$x the regressors. Both are sorted such
that first T observations are those for unit 1, followed by the T
observations for unit 2, etc.}

\item{obs_N, obs_T}{The number of cross-section units (obs_N) and (obs_T) in
the data. Must both be provided.}

\item{TRUNC}{The truncation parameter tau used to estimate the number of
factors. Default is 0.05.}

\item{NFACTORS}{Allows to set the number of factors use in estimation.
Default is NULL so that the data driven choice with TRUNC=0.05 is used.}

\item{variant}{Choose whether to use the lasso based estimator
(variant="Lasso") or the least squares variant (variant="LS"). Must be
specified.}

\item{penalty_choice}{Setting penalty_choice = "None" will ensure that no
data driven lasso penalty selection will be done. Default is to run
cross-validation to select the lasso penalty parameter.}

\item{NFOLDS}{Specify the number of folds to use when using cross-validation
to select lasso penalty based on cv.glmnet. Default is the default in
cv.glmnet, which is 10.}

\item{NLAMBDA}{Specify the number of lasso penalties to use in glmnet when
penalty_choice= "None". Default is the default in glmnet, which is 100.}

\item{scree_plot}{Logical variable to indicate whether a scree plot of the
eigendecomposition of \eqn{\Sigma} should be shown. The default is TRUE.}
}
\value{
In general the function returns the estimated coefficients, the
number of estimated factors and the lasso penalty parameter. More
specifically:
\itemize{
\item Unless penalty_choice ="None", the function returns a list with
\describe{
\item{$coefs}{The coefficient estimates.}
\item{$K_hat}{The estimated number of factors. If NFACTORS was
set by user then this will be returned.}
\item{$Lambda}{The data driven lasso penalty choice. If variant ="LS",
it is set to 0.}
}
\item If the option penalty_choice ="None" was set, then the function
returns a list with
\describe{
\item{$coefs}{The coefficient estimates.}
\item{$K_hat}{The estimated number of factors. If NFACTORS was
set by user then this will be returned.}
\item{$Lambda}{The values of the lasso penalty parameters used. If
variant ="LS", it is set to 0.}
}
}
}
\description{
\code{hd_panel_estimator} fits the high-dimensional CCE estimation procedure
proposed in Vogt, M., Walsh, C., and Linton, 0. (2022) "Estimation of
High-Dimensional Panel Data Models with Interactive Fixed Effects".
}
\examples{
# Load the data set
data("data_example")

# Set the dimensions of the data
obs_N <- 50
obs_T <- 50

# Do the estimation
estimate_model <- hd_panel_estimator(data=data_example,obs_N=obs_N,
                      obs_T=obs_T,TRUNC=0.1,penalty_choice = "None",variant="Lasso")

}
\references{
Vogt, M., Walsh, C. and Linton, O.  (2022) "Estimation of
High-Dimensional Panel Data Models with Interactive Fixed Effects"
}
