% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.r
\name{simulate_data}
\alias{simulate_data}
\title{Simulate a data set with interactive fixed effects}
\usage{
simulate_data(obs_N, obs_T, MU, RHO, p1, p2, p3, K = 3)
}
\arguments{
\item{obs_N}{number of cross-section units.}

\item{obs_T}{number of time periods.}

\item{MU}{vector of mean individual loadings.}

\item{RHO}{pairwise correlation coefficient of the regressors.}

\item{p1}{number of additional regressors that only depend on first factor}

\item{p2}{number of additional regressors that only depend on second factor}

\item{p3}{number of additional regressors that only depend on third factor}

\item{K}{the number of unobserved factors. This has only been tested with
the default value of K=3.}
}
\value{
List containing the balanced panel data with data$y containing
the dependent variables and data$x the regressors. Both are sorted such
that first T observations are those for unit 1, followed by the T
observations for unit 2, etc. Finally, data$f contains the unobserved
factors.
}
\description{
\code{simulate_data} simulates data as in simulation section of Vogt, M.,
Walsh, C., and Linton, 0. (2022) "Estimation of High-Dimensional Panel Data
Models with Interactive Fixed Effects". The example provides the code used
to create the data set \bold{data_example.rda} of the package.
}
\examples{
\dontrun{
# Simulate an example data set as in Scenario B of the paper with
# N = 50, T = 50, rho = 0.25, p = 900
#----------------------------------------------------------------

# Set the parameters
#-------------------

# Set the number of cross-sections, time periods
obs_N  <- 50
obs_T  <- 50
# Specify the number of additional "irrelevant" regressors
p1 <- p2 <- p3 <- 299
# Specify the pairwise correlation among the regressors
rho <- 0.25
# Specify the mean for the factor loadings
MU <- c(rep(1,3),c(diag(0.5,nrow=3)),rep(1,(p1+p2+p3)))

# Simulate the data
set.seed(1048)
data_example <- simulate_data(obs_N=obs_N, obs_T=obs_T, MU=MU, RHO=rho, p1=p1, p2=p2, p3=p3)

# Rearrange the order of the additional "irrelevant" regressors
# for the example data of the package.
#----------------------------------------------------------------

# Rearrange the "additional" columns so that they are not in blocks
indices <- c(1:3)

for (s in 1:p1){
 indices <- c(indices,3+s,3+p1+s,3+2*p1+s)
}

data_example$x <- data_example$x[,indices]

save(data_example, file="data_example.rda",compress=TRUE)
}
}
